myprojectinicial/requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.15.2
psycopg==3.2.5
psycopg-binary==3.2.5
sqlparse==0.5.3
tzdata==2025.1
